pyproject.toml
[build-system]
requires = ["setuptools>=61.0,<75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agp_tool"
version = "0.1.0.2"
description = "Ambulatory glucose profile analysis tool"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
authors = [
    { name = "Darío Clavijo" },
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "matplotlib>=3.4.0",
    "openpyxl>=3.0.0",
    "xlrd>=2.0.1",
    "odfpy>=1.4.1",
    "Pillow>=11.3.0",
    "fpdf2>=2.7.0",
]

[project.urls]
Homepage = "https://github.com/daedalus/agp_tool"
Repository = "https://github.com/daedalus/agp_tool"
"Bug Tracker" = "https://github.com/daedalus/agp_tool/issues"

[project.scripts]
agp = "agp.main:main"

[project.optional-dependencies]
test = ["pytest>=7.0", "pytest-cov"]
dev = ["black>=24.0", "ruff>=0.4.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["agp*"]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]
